requirements.txt
google-auth-oauthlib==1.0.0
google-api-python-client==2.108.0
streamlit==1.29.0
supabase==2.0.3
groq==0.4.1
numpy==1.26.4
pandas==2.0.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.1.0.post1
base58==2.1.1
httpx==0.24.1
python-dateutil==2.8.2
